+++ crates/swc_ecma_transforms_base/src/rename/analyzer/reverse_map.rs
